refactoe: 重构代码结构, 优化api调用
将 Client 的各项功能实现拆分到独立的模块文件中：
- 新增 commit.rs 模块，实现 Commit trait
- 新增 issue.rs 模块，实现 Issue trait
- 新增 org.rs 模块，实现 Org trait
- 新增 repo.rs 模块，实现 Repo trait
- 新增 user.rs 模块，实现 User trait
- 重构 lib.rs，移除原有实现代码，改为导出各模块功能
