+++ nipaw_core/src/lib.rs
@@ -1,10 +1,8 @@
 mod client;
 pub mod error;
 
-pub use client::Client;
+pub use client::{Client, Commit, Issue, Org, Repo, User};
 pub use error::Error;
 pub mod option;
 pub mod types;
-
 pub type Result<T> = std::result::Result<T, Error>;
-pub use types::collaborator::CollaboratorPermission;
